% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/free_ports.R
\name{free_ports}
\alias{free_ports}
\title{Retrieve many (not all) Transmission Control Protocol (TCP) ports that are not currently in use}
\usage{
free_ports()
}
\value{
An integer vector
}
\description{
Retrieve many (not all) Transmission Control Protocol (TCP) ports that are not currently in use
}
\examples{
\dontrun{
  free_ports()
}

}
