% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netstat-package.R
\docType{package}
\name{netstat-package}
\alias{netstat-package}
\title{Retrieve Network Statistics Including Available TCP Ports}
\description{
Retrieve Network Statistics Including Available TCP Ports
}
\details{
It has the goal of providing a convenient way of retrieving and parsing commonly used
   Transmission Control Protocol (TCP) port information provided via the 'netstat' command line
   utility.
}
\author{
Steve Condylios \email{steve.condylios@gmail.com}
}
