% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netstat.R
\name{netstat}
\alias{netstat}
\title{Retrieve the output of the 'netstat' command line utility}
\usage{
netstat()
}
\value{
A character vector containing the lines of the output of the
   'netstat' command line utility
}
\description{
Retrieve the output of the 'netstat' command line utility
}
\examples{
\dontrun{
  netstat()
}

}
\keyword{internal}
