% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ports_in_use.R
\name{ports_in_use}
\alias{ports_in_use}
\title{Obtain the TCP ports currently in use}
\usage{
ports_in_use()
}
\value{
A character vector of the TCP ports currently in use
}
\description{
Obtain the TCP ports currently in use
}
\examples{
\dontrun{
  (active_ports <- ports_in_use())
}

}
