% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/free_port.R
\name{free_port}
\alias{free_port}
\title{Retrieve a Transmission Control Protocol (TCP) port that is not currently in use}
\usage{
free_port(random)
}
\arguments{
\item{random}{Boolean value that defaults to FALSE}
}
\value{
An integer vector of length one
}
\description{
Retrieve a Transmission Control Protocol (TCP) port that is not currently in use
}
\examples{
\dontrun{
  free_port()
  free_port(random = TRUE)
}

}
